requirements.txt
pyrofork
TgCrypto
catbox-uploader
aiohttp
pytz
